requirements-test.txt
apache-airflow==3.0.1
requests>=2.25.0
pyyaml>=5.4.0
pytest>=6.2.5
